\name{Covpath-package}
\alias{Covpath-package}
\alias{Covpath}
\docType{package}
\title{ COVPATH Package}
\description{
 Implements a pathwise algorithm for Covariance Selection
}
\details{
\tabular{ll}{
Package: \tab Covpath\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-03-13\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
List of Functions
covpath.R
covpathCV.R
covpathCVforTS.R
plotDependencyGraph.R
plotDependencyTimeSeries.R
}
\references{
A Pathwise Algorithm for Covariance Selection.
Optimization for Machine Learning, S. Sra, S. Nowozin and S. Wright, editors, MIT Press, 2011. 
}
\author{ Vijay Krishnamurthy, Alexandre d'Aspremont}
\keyword{ multivariate }
