\name{createProtovishtml}
\alias{createProtovishtml}
\title{createProtovishtml}
\description{
Internal Covpath wrapper function to creare html file for display}
\usage{
createProtovishtml(htmlfile, jsfile, graphWidth, graphHeight)
}
\arguments{
  \item{htmlfile}{ html file name}
  \item{jsfile}{ Javascript file name}
  \item{graphWidth}{ Protovis parameter Protovis parameter}
  \item{graphHeight}{  Protovis parameter}
}
\keyword{ multivariate }
