\name{createProtovishtmlforCV}
\alias{createProtovishtmlforCV}
\title{createProtovishtmlforCV}
\description{
Internal Covpath wrapper function to creare html file for display}
\usage{
createProtovishtmlforCV(numPlots, numsamples, dates, htmlfile, jsfile, graphWidth, graphHeight)
}
\arguments{
  \item{numPlots}{ Number of time series plots}
  \item{numsamples}{ Number of samples}
  \item{dates}{Dates defining the series}
  \item{htmlfile}{ html file name}
  \item{jsfile}{ Javascript file name}
  \item{graphWidth}{ Protovis parameter Protovis parameter}
  \item{graphHeight}{  Protovis parameter}
}
\keyword{ multivariate }
