\name{Forwards}
\docType{data}
\alias{Forwards}
\title{Forward Interest Rate data}
\description{
Treasury interest rate data used to demonstrate Covpath functions }
\usage{Forwards}
\format{ cvs file containing forward interest rates.}
\source{ Wharton Research Data Services (WRDS)}
\references{
A Pathwise Algorithm for Covariance Selection.
Optimization for Machine Learning, S. Sra, S. Nowozin and S. Wright, editors, MIT Press, 2011. 
}
\keyword{datasets}