\name{FxData}
\docType{data}
\alias{FxData}
\title{Foreign Exchange Currency data}
\description{
Foreign Exchange Currency data used to demonstrate Covpath functions }
\usage{FxData}
\format{ cvs file containing exchange rates.}
\source{ Wharton Research Data Services (WRDS)}
\references{
A Pathwise Algorithm for Covariance Selection.
Optimization for Machine Learning, S. Sra, S. Nowozin and S. Wright, editors, MIT Press, 2011. 
}
\keyword{datasets}