\name{thresholdAndPrintResults}
\alias{thresholdAndPrintResults}
\title{thresholdAndPrintResults}
\description{
Internal wrapper function to Print results
}
\usage{
thresholdAndPrintResults(Sigma, rholist, XSol, RelGap)
}
\arguments{
  \item{Sigma}{Sample covariance matrix}
  \item{rholist}{Penalty List}
  \item{XSol}{Covariance estimate from covpath}
  \item{RelGap}{Relative improvement from covpath}
}
\keyword{ multivariate }
