\name{thresholdCondCorrMatrix}
\alias{thresholdCondCorrMatrix}
\title{thresholdCondCorrMatrix}
\description{
Internal wrapper function to produce conditional correlation matrix for a particular cardinality
}
\usage{
thresholdCondCorrMatrix(C, thrsQuantile)
}
\arguments{
  \item{C}{ Conditonal correlation matrix from covpath}
  \item{thrsQuantile}{Fraction in [0,1] to specify cardinality}
}
\keyword{ multivariate }
